# Conflicts:
#	client/src/main/java/client/scenes/AdminEventsCtrl.java

+++ client/src/main/java/client/scenes/MainCtrl.java
@@ -17,6 +17,7 @@ package client.scenes;
 
 import client.services.I18N;
 import client.services.NotificationHelper;
+import client.services.NotificationService;
 import client.utils.Config;
 import client.utils.User;
 import commons.Event;
@@ -71,6 +72,7 @@ public class MainCtrl {
 
     private AdminEventsCtrl adminEventsCtrl;
     private Scene adminEvents;
+    private NotificationService notificationService;
 
     public void initialize(Stage primaryStage, Pair<FirstTimeCtrl, Parent> firstTime,
                            Pair<EventOverviewCtrl, Parent> eventOverview,
@@ -157,6 +159,7 @@ public class MainCtrl {
                 }
             }
         });
+        this.notificationService = new NotificationHelper();
     }
 
     public void showInviteView(Event event){
@@ -203,23 +206,21 @@ public class MainCtrl {
     }
 
     public void showLanguageOptions() {
-        primaryStage.setTitle("Splitty: Languages");
+        primaryStage.setTitle(I18N.get("window.settings.language"));
         primaryStage.setScene(languages);
     }
 
-    // TODO Make the actual switch to the languages once button is pressed
     public void switchToEnglish() {
         I18N.setLocale(Locale.ENGLISH);
-        String switchLanguageHeader = I18N.get("language.infoLanguages");
-        String switchLanguageTitle = I18N.get("language.switchTitle");
-        String switchLanguageMessage = I18N.get("language.switchMessage");
-        NotificationHelper notificationHelper = new NotificationHelper();
-        notificationHelper.informUser(switchLanguageTitle, switchLanguageMessage, switchLanguageHeader);
     }
 
     public void switchToDutch() {
         Locale dutch = I18N.getSupportedLocales().get(1);
         I18N.setLocale(dutch);
+    }
+
+    public void uponLanguageSwitch(){
+        primaryStage.setTitle(I18N.get("window.settings.language"));
         String switchLanguageHeader = I18N.get("language.infoLanguages");
         String switchLanguageTitle = I18N.get("language.switchTitle");
         String switchLanguageMessage = I18N.get("language.switchMessage");
@@ -292,6 +293,9 @@ public class MainCtrl {
     }
 
     public void deleteAllData(){
+        if (!notificationService.showConfirmation("Delete data", "Are you sure you want to delete your data? This action cannot be undone.")){
+            return;
+        }
         Config.deleteUserConfigFile();
         this.chooseFirstPage(false);
     }
